tests: restructure tsts in prep for integration test
