+++ .gitignore
@@ -20,4 +20,4 @@
 /CI_reports/
 
 # general temp dir - utilized during tests for reasons
-/.tmp/
+.tmp/
